backend/requirements.txt
# Backend Requirements (Render deployment)
# Heavy dependencies for RAG + FAISS + LangChain

fastapi==0.110.0
uvicorn[standard]==0.27.1
pydantic>=2.7.0
pydantic-settings>=2.0.0

# LangChain with Pydantic v2 compatible versions
langchain>=0.2.0
langchain-community>=0.2.0
langchain-core>=0.2.0
langchain-huggingface>=0.1.0
langchain-text-splitters>=0.2.0
faiss-cpu==1.8.0

# Other dependencies
pyjwt==2.8.0
requests==2.31.0
python-dotenv==1.0.1
tenacity>=8.2.0
huggingface-hub>=0.23.0
pypdf==4.1.0
